homework3/requirements.txt
pandas
requests
scikit-learn
xgboost
mlflow
boto3
sqlalchemy
apache-airflow==3.0.1
pyarrow==20.0.0